+++ tests/org.eclipse.n4js.hlc.tests/src/org/eclipse/n4js/hlc/tests/InstallCompileRunN4jscExternalTest.java
@@ -25,7 +25,7 @@ import org.junit.Test;
 /**
  * Downloads, installs, compiles and runs 'express'.
  */
-public class InstallCompileRunN4jscExternalTest extends BaseN4jscExternalTest {
+public class InstallCompileRunN4jscExternalTest extends AbstractN4jscTest {
 	File workspace;
 
 	/** Prepare workspace. */
