+++ tests/org.eclipse.n4js.hlc.tests/src/org/eclipse/n4js/hlc/tests/InstallCompileRunN4jscExternalWithDefinitionFilesTest.java
@@ -25,7 +25,7 @@ import org.junit.Test;
 /**
  * Downloads, installs, compiles and runs 'express' with N4JS definition file support.
  */
-public class InstallCompileRunN4jscExternalWithDefinitionFilesTest extends BaseN4jscExternalTest {
+public class InstallCompileRunN4jscExternalWithDefinitionFilesTest extends AbstractN4jscTest {
 	File workspace;
 
 	private static final String PROJECT_NAME_N4JS = "project.using.external.from.n4js";
