+++ tests/org.eclipse.n4js.hlc.tests/src/org/eclipse/n4js/hlc/tests/InstallCompileRunN4jscExternalWithSingleProjectCompileTest.java
@@ -27,7 +27,7 @@ import com.google.common.base.Predicates;
 /**
  * Downloads, installs, compiles and runs 'express'.
  */
-public class InstallCompileRunN4jscExternalWithSingleProjectCompileTest extends BaseN4jscExternalTest {
+public class InstallCompileRunN4jscExternalWithSingleProjectCompileTest extends AbstractN4jscTest {
 	File workspace;
 
 	/** Prepare workspace. */
