+++ tests/org.eclipse.n4js.hlc.tests/src/org/eclipse/n4js/hlc/tests/InstallFromManifestCompileRunN4jscExternalImportsTest.java
@@ -29,7 +29,7 @@ import com.google.common.base.Predicates;
  * instructed to discover missing dependencies and to install them before compilation, which is done with
  * {@code --installMissingDependencies} flag.
  */
-public class InstallFromManifestCompileRunN4jscExternalImportsTest extends BaseN4jscExternalTest {
+public class InstallFromManifestCompileRunN4jscExternalImportsTest extends AbstractN4jscTest {
 	File workspace;
 
 	/** Prepare workspace. */
